Move style_gen test
